Relocate validate_parameter function to helpers
